Split categorize into transactions categorize + categories + merchants
CLI v2: categorize workflow (bulk, stats, rules, auto, ml) moves under
transactions categorize *. Category taxonomy becomes top-level 'categories'.
Merchant mappings becomes top-level 'merchants'. Hard cut: old
moneybin categorize * paths removed.
