+++ tests/moneybin/test_cli/test_cli_output_quiet.py
@@ -23,11 +23,11 @@ _READ_ONLY_HELP_PATHS: list[list[str]] = [
     ["import", "history", "--help"],
     ["import", "formats", "list", "--help"],
     ["import", "formats", "show", "--help"],
-    ["categorize", "summary", "--help"],
-    ["categorize", "auto", "stats", "--help"],
-    ["categorize", "auto", "review", "--help"],
-    ["categorize", "auto", "rules", "--help"],
-    ["categorize", "list-rules", "--help"],
+    ["transactions", "categorize", "stats", "--help"],
+    ["transactions", "categorize", "auto", "stats", "--help"],
+    ["transactions", "categorize", "auto", "review", "--help"],
+    ["transactions", "categorize", "auto", "rules", "--help"],
+    ["transactions", "categorize", "rules", "list", "--help"],
     ["sync", "status", "--help"],
     ["profile", "list", "--help"],
     ["profile", "show", "--help"],
