+++ tests/moneybin/test_cli/test_help_no_wizard.py
@@ -29,8 +29,14 @@ _GROUPS = [
     ["profile"],
     ["import"],
     ["sync"],
-    ["categorize"],
+    ["transactions"],
+    ["transactions", "categorize"],
+    ["transactions", "categorize", "rules"],
+    ["transactions", "categorize", "auto"],
+    ["transactions", "categorize", "ml"],
     ["transactions", "matches"],
+    ["categories"],
+    ["merchants"],
     ["transform"],
     ["synthetic"],
     ["track"],
